feat: new impl for token bucket alg(bucket4j) finished script for load testing

+++ src/main/java/com/example/ratelimiter/configuration/ApplicationConfiguration.java
@@ -13,7 +13,7 @@ public class ApplicationConfiguration {
     private final QuotasConfiguration quotasConfiguration;
 
     @Bean
-    public Map<String, Double> getRpsQuotas() {
+    public Map<String, Long> getRpsQuotas() {
         return Collections.unmodifiableMap(quotasConfiguration.getRpsQuotas());
     }
 }
